requirements.txt
rembg==2.0.55
pillow==10.2.0
tqdm==4.66.2
